Modernized RecoTauTag/HLTProducers modules
- removed use of SealModule.cc
- moved plugin registration to individual module .cc files
- changed legacy modules to thread-friendly versions
- minor code improvements
