+++ RecoTauTag/HLTProducers/interface/DQMTauProducer.h
@@ -3,7 +3,7 @@
 
 // user include files
 #include "FWCore/Framework/interface/Frameworkfwd.h"
-#include "FWCore/Framework/interface/EDProducer.h"
+#include "FWCore/Framework/interface/global/EDProducer.h"
 
 #include "FWCore/Framework/interface/Event.h"
 #include "FWCore/ParameterSet/interface/ParameterSet.h"
@@ -13,11 +13,11 @@
 #include "DataFormats/TauReco/interface/L2TauInfoAssociation.h"
 #include "DataFormats/TauReco/interface/HLTTau.h"
 
-class DQMTauProducer : public edm::EDProducer {
+class DQMTauProducer : public edm::global::EDProducer<> {
 public:
   explicit DQMTauProducer(const edm::ParameterSet&);
   ~DQMTauProducer() override;
-  void produce(edm::Event&, const edm::EventSetup&) override;
+  void produce(edm::StreamID, edm::Event&, const edm::EventSetup&) const override;
 
 private:
   edm::EDGetTokenT<reco::IsolatedTauTagInfoCollection> trackIsolatedJets_;
