+++ RecoTauTag/HLTProducers/interface/L1HLTJetsMatching.h
@@ -3,7 +3,7 @@
 
 // user include files
 #include "FWCore/Framework/interface/Frameworkfwd.h"
-#include "FWCore/Framework/interface/EDProducer.h"
+#include "FWCore/Framework/interface/global/EDProducer.h"
 
 #include "FWCore/Framework/interface/Event.h"
 #include "FWCore/ParameterSet/interface/ParameterSet.h"
@@ -17,18 +17,13 @@
 #include <map>
 #include <vector>
 
-class L1HLTJetsMatching : public edm::EDProducer {
+class L1HLTJetsMatching : public edm::global::EDProducer<> {
 public:
   explicit L1HLTJetsMatching(const edm::ParameterSet&);
   ~L1HLTJetsMatching() override;
-  void produce(edm::Event&, const edm::EventSetup&) override;
+  void produce(edm::StreamID, edm::Event&, const edm::EventSetup&) const override;
 
 private:
-  std::vector<l1extra::L1JetParticleRef> tauCandRefVec;
-  std::vector<l1extra::L1JetParticleRef> jetCandRefVec;
-  std::vector<l1extra::L1JetParticleRef> objL1CandRefVec;
-  l1extra::L1JetParticleRef tauCandRef;
-
   edm::EDGetTokenT<edm::View<reco::Candidate> > jetSrc;
   edm::EDGetTokenT<trigger::TriggerFilterObjectWithRefs> tauTrigger;
   double mEt_Min;
