+++ RecoTauTag/HLTProducers/interface/PFJetToCaloProducer.h
@@ -3,7 +3,7 @@
 
 // user include files
 #include "FWCore/Framework/interface/Frameworkfwd.h"
-#include "FWCore/Framework/interface/EDProducer.h"
+#include "FWCore/Framework/interface/global/EDProducer.h"
 
 #include "FWCore/Framework/interface/Event.h"
 #include "FWCore/ParameterSet/interface/ParameterSet.h"
@@ -12,11 +12,11 @@
 #include "DataFormats/JetReco/interface/PFJetCollection.h"
 #include "DataFormats/JetReco/interface/CaloJetCollection.h"
 
-class PFJetToCaloProducer : public edm::EDProducer {
+class PFJetToCaloProducer : public edm::global::EDProducer<> {
 public:
   explicit PFJetToCaloProducer(const edm::ParameterSet&);
   ~PFJetToCaloProducer() override;
-  void produce(edm::Event&, const edm::EventSetup&) override;
+  void produce(edm::StreamID, edm::Event&, const edm::EventSetup&) const override;
 
 private:
   edm::EDGetTokenT<reco::PFJetCollection> tauSrc_;
