+++ RecoTauTag/HLTProducers/interface/PFTauToJetProducer.h
@@ -3,7 +3,7 @@
 
 // user include files
 #include "FWCore/Framework/interface/Frameworkfwd.h"
-#include "FWCore/Framework/interface/EDProducer.h"
+#include "FWCore/Framework/interface/global/EDProducer.h"
 
 #include "FWCore/Framework/interface/Event.h"
 #include "FWCore/ParameterSet/interface/ParameterSet.h"
@@ -13,11 +13,11 @@
 #include "DataFormats/JetReco/interface/CaloJetCollection.h"
 #include "DataFormats/TauReco/interface/PFTau.h"
 
-class PFTauToJetProducer : public edm::EDProducer {
+class PFTauToJetProducer : public edm::global::EDProducer<> {
 public:
   explicit PFTauToJetProducer(const edm::ParameterSet&);
   ~PFTauToJetProducer() override;
-  void produce(edm::Event&, const edm::EventSetup&) override;
+  void produce(edm::StreamID, edm::Event&, const edm::EventSetup&) const override;
 
 private:
   edm::EDGetTokenT<reco::PFTauCollection> tauSrc_;
