+++ RecoTauTag/HLTProducers/src/TauRegionalPixelSeedGenerator.cc
@@ -1,0 +1,19 @@
+// -*- C++ -*-
+//
+// Package:     RecoTauTag/HLTProducers
+// Class  :     TauRegionalPixelSeedGenerator
+//
+// Implementation:
+//     [Notes on implementation]
+//
+// Original Author:  Christopher Jones
+//         Created:  Tue, 13 Sep 2022 21:10:41 GMT
+//
+
+// system include files
+
+// user include files
+#include "TauRegionalPixelSeedGenerator.h"
+#include "RecoTracker/TkTrackingRegions/interface/TrackingRegionProducerFactory.h"
+
+DEFINE_EDM_PLUGIN(TrackingRegionProducerFactory, TauRegionalPixelSeedGenerator, "TauRegionalPixelSeedGenerator");
